requirements.txt
streamlit
fastapi
uvicorn
google-generativeai
python-dotenv
pillow
pymupdf
requests
